# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Adapta API
  description: |
    **Your model. Your data. Your servers.**

    Adapta is a self-hosted model-customization platform. Technical teams deploy
    it on their own infrastructure to specialize and serve private language models two ways,
    each behind an OpenAI-compatible API. Your data never leaves your servers.

    **Give it knowledge (RAG)** — upload documents, embed them, and get answers grounded in
    your content, with citations. Weights never change. CPU-only.

    **Change how it behaves (fine-tuning / LoRA)** — train a QLoRA adapter on instruction
    data; the adapter only serves once it passes the eval gate (score ≥ 0.6 or a clear
    improvement over the base model). Requires a GPU.

    **Composed endpoints** — a single project endpoint can serve both: retrieved context
    with citations (RAG) and trained tone/format (the LoRA adapter) in one call.

    ---

    **Auth:** management endpoints require a JWT Bearer token from `POST /v1/auth/login`.
    The `POST /v1/chat/completions` endpoint uses a scoped project API key (`adp_*` prefix).

  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/yielab/adapta/blob/main/LICENSE
  contact:
    name: Adapta
    url: https://github.com/yielab/adapta
  x-logo:
    url: /console/logo.svg
    altText: Adapta
servers:
  - url: /v1
    description: Versioned API
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an organization + its admin user (open registration)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '422': {$ref: '#/components/responses/Error'}
        '201': {description: User created, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Obtain JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Token, content: {application/json: {schema: {$ref: '#/components/schemas/TokenResponse'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current authenticated user
      security: [{BearerAuth: []}]
      responses:
        '200': {description: User, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
        '401': {$ref: '#/components/responses/Error'}
  /auth/invite:
    post:
      operationId: createInvite
      tags: [auth]
      summary: Invite a user to a team (admin only)
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InviteRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '201': {description: Invitation (token shown once), content: {application/json: {schema: {$ref: '#/components/schemas/InvitationResponse'}}}}
  /auth/accept-invite:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Redeem an invite token and set a password
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptInviteRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '201': {description: User created/joined, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
        '429': {$ref: '#/components/responses/Error'}
  /auth/invitations:
    get:
      operationId: listInvitations
      tags: [auth]
      summary: List a team's invitations (admin only)
      security: [{BearerAuth: []}]
      parameters:
        - name: team_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Invitations, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/InvitationResponse'}}}}}
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password (verifies current password first)
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '204': {description: Password changed}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}

  # ──────────────────────────────────────────────────────────────────────────
  # Teams
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Scoped project API key (adp_...)
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    Role:
      type: string
      enum: [admin, member, viewer]
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, revoked]
...
    RegisterRequest:
      type: object
      required: [org_name, email, password]
      properties:
        # Must contain at least one non-whitespace character — the server
        # rejects blank names (auth.py strips before checking). maxLength
        # values mirror the DB column sizes (orgs.name/users.email) — without
        # them an oversized value overflows the column → 500.
        org_name: {type: string, minLength: 1, maxLength: 128, pattern: "\\S"}
        email: {type: string, format: email, maxLength: 256}
        password: {type: string, minLength: 8}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}

    # Response schemas list every field they always carry in `required`, so the
    # generated models are strict (non-Optional) — the routers construct every
    # field, including the nullable ones. Datetime fields are typed as plain
    # `string` (the handlers emit `.isoformat()` strings), avoiding the
    # aware/naive parsing pitfall a `format: date-time` (AwareDatetime) model
    # would hit on response serialization.
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: {type: string}
        token_type: {type: string, default: bearer}
    UserResponse:
      type: object
      required: [id, email, org_id, teams]
      properties:
        id: {type: string}
        email: {type: string}
        org_id: {type: string}
        teams:
          type: array
          description: Teams the user belongs to, with their role in each. Lets a client discover the team_id required by the project endpoints.
          items:
            $ref: '#/components/schemas/TeamSummary'
    TeamSummary:
      type: object
      required: [id, name, role]
      properties:
        id: {type: string}
        name: {type: string}
        role: {$ref: '#/components/schemas/Role'}

    # Invitations
    InviteRequest:
      type: object
      required: [email, team_id]
      properties:
        email: {type: string, format: email, maxLength: 255}
        team_id: {type: string}
        # Optional on the wire; the handler defaults a missing role to `member`
        # (a $ref + default trips codegen's enum-default mypy bug, so the
        # default lives in code, not the schema).
        role: {$ref: '#/components/schemas/Role'}
    AcceptInviteRequest:
      type: object
      required: [token, password]
      properties:
        token: {type: string, minLength: 1}
        password: {type: string, minLength: 8}
    InvitationResponse:
      type: object
      required: [id, email, team_id, role, status, token, expires_at]
      properties:
        id: {type: string}
        email: {type: string}
        team_id: {type: string}
        role: {$ref: '#/components/schemas/Role'}
        status: {$ref: '#/components/schemas/InvitationStatus'}
        token: {type: string, nullable: true}
        expires_at: {type: string}
...
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: {type: string}
        new_password: {type: string, minLength: 8}

    # Projects
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
        correlation_id: {type: string, nullable: true}
    ErrorResponse:
      type: object
      properties:
        error: {$ref: '#/components/schemas/Error'}